as you search for an international location, a "click here" link appears that takes you to that countries' state department page
omething happens formatting wise, that causes the code pulled from GITHUB to become corrupted  after a push from anyone other than myself. I'm thinking it might be becuase you guys are on a different OS.
I added the StateDeparment (SD) API back, due to Hernan's last push it was overwritten and some erroneous code replaced it when I pulled from git. I learned that the SD API is useless for what we envisioned. so i made take the country from the weather api and create an url to link on our HTML to the state department's page for that country.

[2:53 PM]
things to note about this are as follows: this does not work for contires the SD does not have a page/data for. i.e. the Unitied States of America. additionally... it currently only works with countries with only one name.

[2:54 PM]
in order for it to work with countries with multiply names, i.e. United Kingdom. you have to parse the value of the countries entry with '/' too difficult to tackle as of right now because we already have to JSON.STRINGIFY the results.